Aplicando as alterações
