+++ config.py
@@ -1,0 +1,2 @@
+class Config:
+    SQLALCHEMY_DATABASE_URI = 'sqlite:///items.db'
